+++ src/components/typeahead-search/http/restSearchClient.ts
@@ -65,9 +65,8 @@ export function restSearchClient( getJson: FetchJson = fetchJson ): SearchClient
 		},
 		submitSearchURL(
 			query,
-			submitParams = { title: 'Special:Search' } ) {
-			// eslint-disable-next-line max-len
-			return encodeURI( `/w/index.php?${buildQueryString( { ...submitParams, search: query } )}` );
+			params = { title: 'Special:Search' } ) {
+			return encodeURI( `/w/index.php?${buildQueryString( { ...params, search: query } )}` );
 		}
 	};
 }